Fix API type from int to float

+++ black_scholes_dp.py
@@ -1,9 +1,9 @@
 import numpy as np
 import scipy.stats as si
-
-
 from numpy import float64
-def black_scholes_call_div(S: int, K: int, T: int, r: float, q: float, sigma: float) -> float64:
+
+
+def black_scholes_call_div(S: float, K: float, T: float, r: float, q: float, sigma: float) -> float64:
 
     # S: spot price
     # K: strike price
@@ -21,7 +21,7 @@ def black_scholes_call_div(S: int, K: int, T: int, r: float, q: float, sigma: fl
     return call
 
 
-def black_scholes_put_div(S: int, K: int, T: int, r: float, q: float, sigma: float) -> float64:
+def black_scholes_put_div(S: float, K: float, T: float, r: float, q: float, sigma: float) -> float64:
 
     # S: spot price
     # K: strike price
@@ -39,7 +39,7 @@ def black_scholes_put_div(S: int, K: int, T: int, r: float, q: float, sigma: flo
     return put
 
 
-def euro_vanilla_dividend(S: int, K: int, T: int, r: float, q: float, sigma: float, option: str = 'call') -> float64:
+def euro_vanilla_dividend(S: float, K: float, T: float, r: float, q: float, sigma: float, option: str = 'call') -> float64:
 
     # S: spot price
     # K: strike price
